pyproject.toml
[project]
name = "fora"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "gunicorn>=22.0.0",
    "whitenoise>=6.5.0",
    "google-genai>=1.64.0",
    "anthropic>=0.82.0",
]
